requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
crewai==0.28.0
openai>=1.13.3,<2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0
selenium==4.15.2
webdriver-manager==4.0.1
langchain-openai==0.0.5
pdfminer.six>=20221105
python-pptx>=0.6.21
python-docx>=0.8.11 